+++ changelog.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Known Issues ##
+- Solution console: The color of the input text is not working properly for the first command
+
+## [1.1.0] - 2018-11-05 ##
+### Added ###
+- Solution console
+
 ## [1.0.1] - 2018-11-04 ##
 ### Fixed ###
 - readme.md image tags (case for file names and the alt texts)
@@ -13,6 +20,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release
 - Solution (Project) Templates functionality
 
-
+[1.1.0]: https://github.com/adamecr/RadProjectsExtension/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/adamecr/RadProjectsExtension/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/adamecr/RadProjectsExtension/releases/tag/v1.0.0